backend/requirements.txt
scenedetect[opencv-headless]>=0.6.2
openai>=1.0.0
Pillow>=10.0.0
fal-client>=0.5.0
requests>=2.31.0
google-genai>=1.0.0
pytest>=7.0.0
